Add a new feature to game() function
In the previous commit, in game() function the winner of the game was
announced after the user and the computer had played 5 rounds. But
sometimes there is no need to play 5 rounds to be decided who the winner
is. If, for example, one of the two opponents wins by 3-1 in the fourth
round, there is no need to play a fifth round. And the new version of
the code fixes just that: Breaks out of the loop when we have a winner
after having already announced the winner.

Also, this new version gets rid of the 'Testing computerPlay()'
section of code at the bottom of the document, as there is no need for
it.