Delete and modify functions
I am about to make a UI for Rock Paper Scissors game (RPS game, from now
 on). In order to do that, first I had to make some changes
to the original code.

* In the console version of the RPS game a function is needed to
capitalize the first letter of the user input.
In the new UI version we don't need to do that because the user is
going to click a button to enter his input. That's why we don't need
that function anymore. Thus, function deleted.

* In the console version of the game we should also check that the user
enters the right input. That is achieved with a number of if blocks
within the functions.
In the UI version that is not needed anymore. Again, entering input is
done by clicking buttons. For this reason the if blocks are deleted as
well.

* Lastly, in the console version user and computer play exactly 5 rounds.
He who wins more rounds, wins game.
In the UI version he who is first to reach 5 wins, wins game. The number
of rounds to play in this case is unlimited.
Therefore, I had to modify the game function accordingly.