Add style.css and reset.css, modify index.html, script.js
* index.html
I added a container with a 'h1' header and a button.

* style.css
Added some basic style while I am working on the script.js

* script.js
I commented out the previous version of the code and started working on
the graphical UI of the game. So far, I have added some basic
interactivity to it. More is to come.