Add variables, add buttons, modify function
Once the game is finished and the winner is decided, I want to ask user
the question if he wants to play another game. if user wants to play
another game, they will have to click on the `Yes` button, otherwise
click on `No`. If user click on `Yes`, game starts over. If user clicks
on `No`, `Game Over` is printed.

To do the above actions, I added new variables modified function
startOver so that the question and two new buttons (`Yes` and `No`) will
appear on the browser and added event listeners to those buttons.